= Extract preprocessor
